+++ src/Sagas/index.js
@@ -1,4 +1,4 @@
-import mergeApiSaga from './mergeApiDataSaga';
+import mergeApiSaga from './mergeApiData';
 
 export default function* rootSaga() {
   yield mergeApiSaga();
