+++ src/Sagas/mergeApiData/index.js
@@ -1,13 +1,15 @@
 import { put, select, takeLatest } from 'redux-saga/effects';
 import { sample } from 'lodash';
-import nameSaga from './nameSaga';
-import dogSaga from './dogSaga';
-import { addDogResults, changeLoading } from '../Redux/Dog/DogRedux';
+import nameSaga from '../name';
+import dogSaga from '../dog';
+import { addDogResults, changeLoading } from '../../Redux/dog/index';
 
 function* fetchApiData() {
   yield nameSaga();
   yield dogSaga();
 }
+
+// *Since I am combining the data from API this is to serve as the combiner
 
 function* workMergeApiDatas() {
   const dogsData = yield select((state) => state.reducer.dog.results);
